refactored LevelHandler to use functions not class
Using a class meant that the app was cluttered with single-use
instances of LevelHandler which menat that the file was also being
opened and left open multiple times.

By using a generator it is now able to open and use the file only
when necessary and while this may result in more reads, this means
that the file is safer and leaves less clutter in memory.