refactored find_by_entity to use type rather than instance
Function shouldn't be taking an instance to search with as this
means instantiating agents with additional arguments and is not
safe or clean to implement. By using the typing.Type it allows
for more generic functions that require less internal class
access and only provide enough for the comparison to take place.