converted find_all_paths from dfs to bfs
Using dfs was an unoptimal solution for my use case and needs.
I required a set of the shortest paths which the agents could choose
from and instead, I was being given all paths regardless of how optimal.
Due to the fact that dfs would also provide some of the most suboptimal
paths first, it was also hard to prune these paths during iteration
and so this became a very slow operation.

Switching to bfs allowed me to find the most optimal paths first, and then
only retain the first 5 paths which can then be provided to the agent.
This number is dynamic and can be changed later if needed for fine-tuning.