refactored test due to optmisation to path finding
When the path finding function was refactored to use bfs, it started
to prioritise the shortest paths. This means that when the list of
paths is returned, it would return in ascending order of path length.
The test was expecting the paths in the wrong order and so this had
to be refactored so that the expected value was in ascending order
too.