refactored entity node search to not use Agent
In an attempt to decouple the agent from the graph, I am rewriting
functions which use the agent to use alternative logic. The aim
is to avoid the need for the graph to depend on agent. This way,
I can use the graph within Agent as it should be used and this way
it reduces the likelihood of a circular import.