refactored move_agent to reduce dependency on Agent imports
There is now less conditional handling taking place within the
move function. This logic should not have been there in the first
place and now, if a collision between two "in-play" entities, a
CollisionException is raised so that the logic can be passed back
into the GameManager to deal with.