Removed type length check within find_node_by_entity
This check was unnecessary as the graph construction logic should not
happen here. It should not be up to this method to police how the graph
is constructed. If the graph is invalid this should be checked during
conversion.

Removing this logic also helps to reduce the circular import between the
Graph and Agent objects.