Removed Agent reference from collision check
It does not need to be checked whether the moving entity is an agent, it
can be assumed from the logic above this snippet as well as the previous
commits justification; the method can only be called in relation to
agents and therefore it does not need to be checked. This also helps to
alleviate the circular import problem.