Refactored Node entities
Node now contains a list to represent entities instead of a single
entity. This is done because in the previous format, ghosts could not be
in the same space as a pickup without a collision issue. This allows
ghosts and pickup items to exist in the same space while handling all
necessary collisions.