Added case for multi-agent nodes
In a node where two agents exist, their priority was previously not
defined, this caused issues when raising the exception as it was
unclear.

The higher priority agent is now the most recently added agent, this
means the agent which has moved to collide with the agent already in the
node, now seen as the lowest priority.

In exceptions, this now means that the collided agent is treated the
same as a pickup item and the logic is maintained regardless of what is
collided with.