Removed lock on max number of entities in Node
Currently the limit is two. The issue here is that, if Pac-Man moves
into a space which contains a pickup, and then the ghost tries to move
into that space too, then Pac-Man is immune from dying as the exception
stops the ghost from moving. By removing this, the ghost can consume
Pac-Man as expected.