Implemented very rudimentary mind for Pac-Man Agent.
The agent's mind is intended to randomly choose a location on the graph
and travel to that location. If, at any point, the path becomes unsafe,
Pac-Man should pick a new location and travel to that location instead.

This is designed to be the first iteration of Pac-Man's mind and is
considered the most basic. It will be improved on with iterative design.